chore: upgrade to TypeScript 7, ESLint 10, and Next.js 16
Combines the three major updates from the Renovate dashboard (#1539).

- next, @next/mdx, eslint-config-next 15.5.25 -> 16.3.4
- eslint 9 -> 10; drop @eslint/eslintrc + @eslint/js (FlatCompat no longer
  needed, eslint-config-next 16 exports flat configs). Add @eslint/compat
  fixupConfigRules because eslint-plugin-react/-import/-jsx-a11y still use
  context methods removed in ESLint 10.
- typescript 5.9 -> 7.0 via the side-by-side setup: typescript aliased to
  @typescript/typescript6 (JS API for typescript-eslint and next build),
  @typescript/native aliased to typescript@7 (native tsc for pnpm typecheck).
- tsconfig: explicit types list (TS 6 defaults it to []), jsx react-jsx and
  .next/dev/types include written by Next 16.
- Move the custom TOC remark plugin and the rehype-autolink-headings wrapper
  into src/mdx-plugins/ and reference them by absolute path so the MDX loader
  options are serializable for Turbopack.
- revalidateTag now requires a cache profile; use { expire: 0 } to keep the
  previous expire-immediately behaviour.
