+++ package.json
@@ -25,7 +25,7 @@
 	"dependencies": {
 		"@imgix/js-core": "^3.8.0",
 		"@mdx-js/loader": "^3.1.1",
-		"@next/mdx": "15.5.25",
+		"@next/mdx": "16.3.4",
 		"@sindresorhus/slugify": "^3.0.1",
 		"@types/mdx": "^2.0.14",
 		"airtable": "^0.12.2",
@@ -39,7 +39,7 @@
 		"leaflet": "^1.9.4",
 		"luxon": "^3.7.2",
 		"mdast-util-toc": "^7.1.0",
-		"next": "15.5.25",
+		"next": "16.3.4",
 		"react": "^19.2.8",
 		"react-dom": "^19.2.8",
 		"react-leaflet": "^5.0.0",
@@ -55,8 +55,7 @@
 		"unified": "^11.0.5"
 	},
 	"devDependencies": {
-		"@eslint/eslintrc": "^3.3.7",
-		"@eslint/js": "^9.39.5",
+		"@eslint/compat": "^2.1.1",
 		"@faker-js/faker": "^10.6.0",
 		"@types/leaflet": "^1.9.22",
 		"@types/leaflet.markercluster": "^1.5.6",
@@ -65,16 +64,17 @@
 		"@types/react": "^19.2.18",
 		"@types/react-dom": "^19.2.5",
 		"@types/sanitize-html": "^2.16.1",
+		"@typescript/native": "npm:typescript@^7.0.2",
 		"concurrently": "^10.0.5",
 		"cross-env": "^10.1.0",
-		"eslint": "^9.39.5",
-		"eslint-config-next": "15.5.25",
+		"eslint": "^10.9.1",
+		"eslint-config-next": "16.3.4",
 		"netlify-cli": "^27.4.2",
 		"npm-watch": "^0.13.0",
 		"prettier": "^3.9.6",
 		"sass": "^1.103.1",
 		"tsx": "^4.23.13",
-		"typescript": "^5.9.3"
+		"typescript": "npm:@typescript/typescript6@^6.0.2"
 	},
 	"overrides": {}
 }
